+++ packages/augur-ui/src/modules/common-elements/chevron-flip.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import classNames from "classnames";
 import PropTypes from "prop-types";
 
-import Styles from "modules/common/components/chevron-flip/chevron-flip.styles";
+import Styles from "modules/common-elements/chevron-flip.styles";
 
 const ChevronFlipIcon = (className = "", fillColor = "#A7A2B2") => (
   <svg className={className} viewBox="0 0 16 16">
