+++ packages/augur-ui/src/modules/common-elements/confirm-table.less
