+++ packages/augur-ui/src/modules/common-elements/form.styles.less
@@ -71,3 +71,264 @@
     width: @size-24;
   }
 }
+
+
+.Input {
+  .border(true, @border-all, @color-input-border-no-focus, 1px);
+
+  background-color: @color-input-background;
+  display: flex;
+
+  button,
+  .input-value-comparison {
+    background-color: @color-input-background;
+    font-weight: @weight-medium;
+  }
+
+  input {
+    .border(true, @border-all, @color-transparent, 1px);
+
+    display: flex;
+    flex: 1;
+  }
+
+  input::placeholder {
+    color: @color-input-border-no-focus;
+  }
+
+  input:focus {
+    .border(true, @border-all, @color-transparent, 1px);
+  }
+
+  &:focus-within {
+    .border(true, @border-all, @color-input-border, 1px);
+  }
+
+  .box {
+    font-size: inherit;
+    padding: @size-16;
+    width: 100%;
+
+    &.search-input {
+      padding-left: 0;
+    }
+  }
+
+  .close {
+    background-color: @color-transparent;
+    height: @size-16;
+    line-height: @size-16;
+    width: @size-16;
+
+    > svg > path {
+      stroke: @color-primary-text;
+    }
+
+    @media @breakpoint-mobile {
+      max-height: @size-24;
+      min-height: @size-24;
+      min-width: @size-24;
+    }
+  }
+}
+
+.noFocus {
+  &:focus-within {
+    .border(true, @border-all, @color-transparent, 1px);
+  }
+}
+
+.Max {
+  .text-10-medium;
+
+  color: @color-neutral;
+  padding: 0 @size-16;
+  text-transform: capitalize;
+}
+
+.MaxDark {
+  background-color: @color-lightergray !important;
+}
+
+.Incremental {
+  .border(true, @border-all, @color-transparent, 1px);
+
+  background-color: @color-input-background;
+  display: flex;
+
+  .box {
+    font-size: inherit;
+    padding: @size-16;
+    width: 80%;
+
+    &.search-input {
+      padding-left: 0;
+    }
+  }
+
+  button,
+  .ValueComparison {
+    background-color: @color-input-background;
+  }
+
+  input {
+    .border(true, @border-all, @color-transparent, 1px);
+
+    display: flex;
+    flex: 1 auto;
+  }
+
+  input::placeholder {
+    color: @color-input-border-no-focus;
+  }
+
+  input:focus {
+    .border(true, @border-all, @color-transparent, 1px);
+  }
+
+  &:focus-within {
+    .border(true, @border-all, @color-input-border, 1px);
+  }
+
+  input[type='number'] {
+    appearance: textfield;
+  }
+
+  input[type=number]::-webkit-outer-spin-button,
+  input[type=number]::-webkit-inner-spin-button {
+    appearance: none;
+    margin: 0;
+  }
+}
+
+.ValueIncrementers {
+  // .border(true, @border-left);
+
+  display: flex;
+  flex-direction: column;
+
+  button {
+    cursor: default;
+    flex: 1;
+    padding: 0 @size-4;
+
+    &:active {
+      background-color: @color-gray;
+    }
+  }
+}
+
+.FocusBorder {
+  .border(true, @border-all, @color-input-border, @border-width-2);
+}
+
+.LightBorder {
+  .border(true, @border-all, @color-lightgray, 1px);
+
+  &:focus-within {
+    .border(true, @border-all, @color-lightgray, 1px);
+  }
+}
+
+.SetWidth {
+  min-width: 17.81rem;
+}
+
+
+
+.InputDropdown {
+  background-color: @color-input-background;
+  color: @color-lightgray;
+  cursor: pointer;
+  display: flex;
+  padding: @size-8 @size-16;
+  position: relative;
+  width: 100%;
+  z-index: @default-z-index;
+
+  &:focus {
+    outline: none;
+  }
+
+  > span:first-of-type {
+    .text-18;
+  
+    line-height: @size-24;
+    padding: 0;
+    text-align: left;
+    width: 100%;
+  
+    &.selected {
+      color: @color-darkpurple;
+    }
+  }
+
+  > div {
+    .text-18;
+  
+    background-color: @color-secondary-text;
+    box-shadow: 0 2px @size-4 rgba(0, 0, 0, 0.5);
+    color: @color-darkpurple;
+    display: none;
+    line-height: @size-24;
+    max-height: 10rem;
+    overflow: auto;
+    padding: @size-8 0;
+    position: absolute;
+    right: 0;
+    top: 100%;
+    width: 100%;
+  
+    &.active {
+      display: block;
+    }
+  
+    > button {
+      background-color: inherit;
+      cursor: pointer;
+      padding: @size-4 @size-16;
+      text-align: left;
+      transition: background-color 0.15s ease;
+      white-space: nowrap;
+      width: 100%;
+  
+      &:hover,
+      &.active {
+        background-color: @color-lightergray;
+      }
+    }
+  }
+
+  > select {
+    display: none;
+    width: 100%;
+  
+    &.selected {
+      color: @color-darkpurple;
+    }
+  }
+
+  > span:last-of-type {
+    align-self: flex-end;
+  
+    svg {
+      height: @size-16;
+  
+      polyline {
+        stroke: @color-lightgray;
+      }
+    }
+  }
+
+  @media @breakpoint-mobile-small {
+    > span:first-of-type,
+    > div,
+    > div.active {
+      display: none;
+    }
+  
+    > select {
+      display: block;
+    }
+  }
+}
