+++ packages/augur-ui/src/modules/common-elements/paginator_v1.tsx
@@ -1,3 +1,5 @@
+// TODO: use only the new pagination going forward, this is depreciated.
+
 import React, { Component } from "react";
 import PropTypes from "prop-types";
 import { Link } from "react-router-dom";
@@ -7,7 +9,7 @@ import { PAGINATION_PARAM_NAME } from "modules/routes/constants/param-names";
 import parseQuery from "modules/routes/helpers/parse-query";
 import makeQuery from "modules/routes/helpers/make-query";
 
-import Styles from "modules/common/components/paginator/paginator.styles";
+import Styles from "modules/common-elements/paginator_v1.styles";
 
 interface PaginatorProps {
   itemsLength: number;
