--- packages/augur-ui/src/modules/common/components/input-dropdown/input-dropdown.styles.less
@@ -1,98 +1,0 @@
-@import "~assets/styles/shared.less";
-
-.InputDropdown {
-  background-color: @color-input-background;
-  color: @color-lightgray;
-  cursor: pointer;
-  display: flex;
-  padding: @size-8 @size-16;
-  position: relative;
-  width: 100%;
-  z-index: @default-z-index;
-
-  &:focus {
-    outline: none;
-  }
-}
-
-.InputDropdown__label {
-  .text-18;
-
-  line-height: @size-24;
-  padding: 0;
-  text-align: left;
-  width: 100%;
-
-  &.selected {
-    color: @color-darkpurple;
-  }
-}
-
-.InputDropdown__list {
-  .text-18;
-
-  background-color: @color-secondary-text;
-  box-shadow: 0 2px @size-4 rgba(0, 0, 0, 0.5);
-  color: @color-darkpurple;
-  display: none;
-  line-height: @size-24;
-  max-height: 10rem;
-  overflow: auto;
-  padding: @size-8 0;
-  position: absolute;
-  right: 0;
-  top: 100%;
-  width: 100%;
-
-  &.active {
-    display: block;
-  }
-
-  > button {
-    background-color: inherit;
-    cursor: pointer;
-    padding: @size-4 @size-16;
-    text-align: left;
-    transition: background-color 0.15s ease;
-    white-space: nowrap;
-    width: 100%;
-
-    &:hover,
-    &.active {
-      background-color: @color-lightergray;
-    }
-  }
-}
-
-.InputDropdown__select {
-  display: none;
-  width: 100%;
-
-  &.selected {
-    color: @color-darkpurple;
-  }
-}
-
-.InputDropdown__icon {
-  align-self: flex-end;
-
-  svg {
-    height: @size-16;
-
-    polyline {
-      stroke: @color-lightgray;
-    }
-  }
-}
-
-@media @breakpoint-mobile-small {
-  .InputDropdown__label,
-  .InputDropdown__list,
-  .InputDropdown__list.active {
-    display: none;
-  }
-
-  .InputDropdown__select {
-    display: block;
-  }
-}
