--- packages/augur-ui/src/modules/common/components/input/input.styles.less
@@ -1,162 +1,0 @@
-@import (reference) '~assets/styles/shared';
-
-.Input {
-  .border(true, @border-all, @color-input-border-no-focus, 1px);
-
-  background-color: @color-input-background;
-  display: flex;
-
-  button,
-  .input-value-comparison {
-    background-color: @color-input-background;
-    font-weight: @weight-medium;
-  }
-
-  input {
-    .border(true, @border-all, @color-transparent, 1px);
-
-    display: flex;
-    flex: 1;
-  }
-
-  input::placeholder {
-    color: @color-input-border-no-focus;
-  }
-
-  input:focus {
-    .border(true, @border-all, @color-transparent, 1px);
-  }
-
-  &:focus-within {
-    .border(true, @border-all, @color-input-border, 1px);
-  }
-
-  .box {
-    font-size: inherit;
-    padding: @size-16;
-    width: 100%;
-
-    &.search-input {
-      padding-left: 0;
-    }
-  }
-
-  .close {
-    background-color: @color-transparent;
-    height: @size-16;
-    line-height: @size-16;
-    width: @size-16;
-
-    > svg > path {
-      stroke: @color-primary-text;
-    }
-
-    @media @breakpoint-mobile {
-      max-height: @size-24;
-      min-height: @size-24;
-      min-width: @size-24;
-    }
-  }
-}
-
-.noFocus {
-  &:focus-within {
-    .border(true, @border-all, @color-transparent, 1px);
-  }
-}
-
-.Input__max {
-  .text-10-medium;
-
-  color: @color-neutral;
-  padding: 0 @size-16;
-  text-transform: capitalize;
-}
-
-.Input__maxDark {
-  background-color: @color-lightergray !important;
-}
-
-.Input__Incremental {
-  .border(true, @border-all, @color-transparent, 1px);
-
-  background-color: @color-input-background;
-  display: flex;
-
-  .box {
-    font-size: inherit;
-    padding: @size-16;
-    width: 80%;
-
-    &.search-input {
-      padding-left: 0;
-    }
-  }
-
-  button,
-  .input-value-comparison {
-    background-color: @color-input-background;
-  }
-
-  input {
-    .border(true, @border-all, @color-transparent, 1px);
-
-    display: flex;
-    flex: 1 auto;
-  }
-
-  input::placeholder {
-    color: @color-input-border-no-focus;
-  }
-
-  input:focus {
-    .border(true, @border-all, @color-transparent, 1px);
-  }
-
-  &:focus-within {
-    .border(true, @border-all, @color-input-border, 1px);
-  }
-
-  input[type='number'] {
-    appearance: textfield;
-  }
-
-  input[type=number]::-webkit-outer-spin-button,
-  input[type=number]::-webkit-inner-spin-button {
-    appearance: none;
-    margin: 0;
-  }
-}
-
-.value__incrementers {
-  // .border(true, @border-left);
-
-  display: flex;
-  flex-direction: column;
-
-  button {
-    cursor: default;
-    flex: 1;
-    padding: 0 @size-4;
-
-    &:active {
-      background-color: @color-gray;
-    }
-  }
-}
-
-.focusBorder {
-  .border(true, @border-all, @color-input-border, @border-width-2);
-}
-
-.lightBorder {
-  .border(true, @border-all, @color-lightgray, 1px);
-
-  &:focus-within {
-    .border(true, @border-all, @color-lightgray, 1px);
-  }
-}
-
-.setWidth {
-  min-width: 17.81rem;
-}
