+++ packages/augur-ui/src/modules/create-market/components/create-market-form-time/create-market-form-time.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import PropTypes from "prop-types";
 import "react-dates/initialize";
 import "react-dates/lib/css/_datepicker.css";
-import InputDropdown from "modules/common/components/input-dropdown/input-dropdown";
+import { InputDropdown } from "modules/common-elements/form";
 import { SingleDatePicker } from "react-dates";
 import {
   ChevronLeft,
