+++ packages/augur-ui/src/modules/create-market/components/create-market-form-time/market-create-time-display.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import classNames from "classnames";
 import PropTypes from "prop-types";
 import ReactTooltip from "react-tooltip";
-import TooltipStyles from "modules/common/less/tooltip.styles";
+import TooltipStyles from "modules/common-elements/tooltip.styles";
 import { Hint } from "modules/common-elements/icons";
 import Styles from "modules/create-market/components/create-market-form-define/create-market-form-define.styles";
