+++ packages/augur-ui/src/modules/forking/components/migrate-rep-confirm/migrate-rep-confirm.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import PropTypes from "prop-types";
 
-import ConfirmStyles from "modules/common/less/confirm-table";
+import ConfirmStyles from "modules/common-elements/confirm-table";
 
 const MigrateRepConfirm = ({ selectedOutcomeName, repAmount, gasEstimate }) => (
   <article className={ConfirmStyles.Confirm}>
