+++ packages/augur-ui/src/modules/market/components/market-properties/market-properties.tsx
@@ -17,7 +17,7 @@ import { ValueDenomination } from "modules/common-elements/labels";
 import getValue from "utils/get-value";
 import { dateHasPassed } from "utils/format-date";
 import Styles from "modules/market/components/market-properties/market-properties.styles";
-import ChevronFlip from "modules/common/components/chevron-flip/chevron-flip";
+import ChevronFlip from "modules/common-elements/chevron-flip";
 import { constants } from "services/augurjs";
 import { FavoritesButton } from "modules/common-elements/buttons";
