+++ packages/augur-ui/src/modules/markets-list/components/markets-list.tsx
@@ -2,7 +2,7 @@ import React, { Component } from "react";
 
 import { PAGINATION_PARAM_NAME } from "modules/routes/constants/param-names";
 import MarketPreview from "modules/market/containers/market-preview";
-import Paginator from "modules/common/components/paginator/paginator";
+import Paginator from "modules/common-elements/paginator_v1";
 import NullStateMessage from "modules/common-elements/null-state-message";
 import { TYPE_TRADE } from "modules/common-elements/constants";
 import isEqual from "lodash/isEqual";
