+++ packages/augur-ui/src/modules/modal/components/modal-network-connect.tsx
@@ -2,7 +2,7 @@ import React, { Component } from "react";
 import classNames from "classnames";
 
 import { InputErrorIcon } from "modules/common-elements/icons";
-import Input from "modules/common/components/input/input";
+import { Input } from "modules/common-elements/form";
 import Styles from "modules/modal/components/common/common.styles.less";
 import ModalActions from "modules/modal/components/common/modal-actions";
 import { windowRef } from "utils/window-ref";
