+++ packages/augur-ui/src/modules/reporting/components/common/disputing-markets.tsx
@@ -3,7 +3,7 @@ import PropTypes from "prop-types";
 
 import NullStateMessage from "modules/common-elements/null-state-message";
 import DisputeMarketCard from "modules/reporting/components/dispute-market-card/dispute-market-card";
-import Paginator from "modules/common/components/paginator/paginator";
+import Paginator from "modules/common-elements/paginator_v1";
 import MarketsHeaderLabel from "modules/markets-list/components/markets-header-label/markets-header-label";
 
 export default class DisputingMarkets extends Component {
