+++ packages/augur-ui/src/modules/reporting/components/reporting-dispute-confirm/reporting-dispute-confirm.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import PropTypes from "prop-types";
 
-import ConfirmStyles from "modules/common/less/confirm-table";
+import ConfirmStyles from "modules/common-elements/confirm-table";
 
 const ReportingDisputeConfirm = ({
   isMarketInValid,
