+++ packages/augur-ui/src/modules/reporting/components/reporting-payouts/reporting-payouts.tsx
@@ -4,7 +4,7 @@ import classNames from "classnames";
 import ReactTooltip from "react-tooltip";
 import { formatAttoRep } from "utils/format-number";
 import Styles from "modules/reporting/components/reporting-payouts/reporting-payouts.styles";
-import TooltipStyles from "modules/common/less/tooltip.styles";
+import TooltipStyles from "modules/common-elements/tooltip.styles";
 import { InputErrorIcon } from "modules/common-elements/icons";
 import { MALFORMED_OUTCOME } from "modules/common-elements/constants";
